Ahora se añade automaticamente al creador del evento como asistente
