+++ controllers/UsuariosController.php
@@ -244,15 +244,4 @@ class UsuariosController extends Controller
     {
         return Yii::$app->user->id === 1 || Yii::$app->user->id === $model->id;
     }
-
-    public function asistire($usuario, $evento)
-    {
-        if (!Yii::$app->user->isGuest) {
-            $sql = 'insert into usuarios_eventos(usuario_id, evento_id) values(' . $usuario->id . ', ' . $evento->id . ')';
-            Yii::$app()->db->createCommand($sql)->execute();
-            Yii::$app->session->setFlash('info', 'Te has añadido satisfactoriamente como asistente');
-        } else {
-            Yii::$app->session->setFlash('error', 'Debes estar logeado para marcarte como asistente');
-        }
-    }
 }
